break out traits
